requirements.txt
selenium==4.35.0
openai==2.15.0
python-dotenv==1.2.1